Rename file PlyrComponent -> Plyr for consistency
