+++ .typo-ci.yml
@@ -21,5 +21,5 @@ excluded_files:
 # Any typos we should ignore?
 excluded_words:
   - typoci
-  - PlyrComponent
+  - Plyr
   - npmrc
